Don't fade into ObsStatus in MyObservations
