+++ src/components/ObservationsFlashList/ObsUploadStatus.tsx
@@ -56,9 +56,7 @@ const ObsUploadStatus = ( {
       progress={progress}
       uniqueKey={observation.uuid}
       queued={queued}
-    >
-      {obsStatus}
-    </UploadStatus>
+    />
   );
 };
